+++ fair/lib/src/experiment/sugar.dart
@@ -43,7 +43,23 @@ class Sugar {
     return data.mapEach((index, item) => builder(item));
   }
 
+  static K switchCase<T, K>(T key,List<SugarSwitchCaseObj<T,K>> caseObjects ,K defaultValue){
+    for (SugarSwitchCaseObj<T,K> sugarCase in caseObjects){
+      if(sugarCase.sugarCase == key){
+        return sugarCase.reValue;
+      }
+    }
+    return defaultValue;
+  }
+
   static Color colorWithOpacity = Colors.grey.withOpacity(0.8);
+
+  static Color colorsWithOpacity(Color c,double o){
+    return c.withOpacity(o);
+  }
+  static String convertToString<T>({required T orginalValue}){
+    return orginalValue.toString();
+  }
 
   static double height(BuildContext context) => MediaQuery.of(context).size.height;
 
@@ -59,3 +75,11 @@ class Sugar {
 
   static Null Function() onTapEmpty() => () {};
 }
+class SugarSwitchCaseObj<T,K>{
+
+  final K reValue;
+  final T sugarCase;
+
+  SugarSwitchCaseObj({required this.sugarCase,required this.reValue});
+}
+