+++ annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.3.0
+* Add null-safety
+* @FairPatch annotate add 'module' parameter
+
 ## 0.2.0
 * @FairWell annotate a filed for bundle generation
 * @FairPatch annotate a Widget as FairWidget