+++ compiler/CHANGELOG.md
@@ -1,5 +1,14 @@
+## [1.1.0]
+* Support singleton parsing.
+
+## [1.0.1]
+* bugfix.
+
+## [1.0.0]
+* Add null-safety.
+
 ## [0.3.0+1]
-* bug fix
+* bugfix.
 
 ## [0.3.0]
 * Minimize size of assets.