+++ dart2dsl/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- bugfix.
+
 ## 1.0.0
 
-- Initial version, created by Stagehand
+- Initial version, created by Stagehand.