+++ fair/CHANGELOG.md
@@ -1,3 +1,19 @@
+## [2.5.0]
+* Add support of flutter 2.8.0、2.8.1.
+* Support singleton parsing.
+* Add Sugar：Sugar.colorsWithOpacity、Sugar.switchCase、Sugar.convertToString.
+
+## [2.4.1]
+* Fix FairLogger import problem.
+* Upgrade Analyzer to 2.3.0.
+
+## [2.4.0]
+* FlatBuffers supports generating bin files in a null safe environment.
+
+## [2.3.0]
+* Add null-safety
+* Add support of flutter 2.5.0、2.5.1、2.5.2、2.5.3
+
 ## [2.0.0]
 * Bug fix of fair_version widgets
 * Add support of flutter 2.0.6